refactor: ImageFile 패키지 이동 및 변수 접근 제어자 변경
