+++ back/src/main/java/com/baba/back/common/FileHandler.java
@@ -1,6 +1,6 @@
 package com.baba.back.common;
 
-import com.baba.back.content.domain.ImageFile;
+import com.baba.back.content.domain.content.ImageFile;
 
 public interface FileHandler {
     /**
